Extract firmware/clean fan-out into clean.py
Seventh structural-split PR for controllers/firmware/controller.py
(follows #733 persistence + #735 cli + #739 lifecycle + #740
factories + #745 runner + #746 follow). Pulls the firmware/clean
fan-out unit — the clean public command plus its two private
helpers — into a new clean.py submodule.

The unit is self-contained: the only caller of
_fan_out_clean_to_connected_peers and _active_build_for is the
clean WS command itself, and the only caller of clean is the WS
dispatch table (via the @api_command decorator). Tests exercise
the public clean entry point and don't monkey-patch the
private helpers — so the private bound methods can be removed
entirely rather than kept as delegates.

Changes:

* New clean.py (185 lines) — three free functions plus the
  _BUILD_PRODUCING_JOB_TYPES module-level constant (the
  constant was only consumed by _active_build_for, so it moves
  with its consumer). The free functions call each other
  directly (peer-function calls) since no test patches them.
* controller.py — 1037 -> 877 lines (-160). The @api_command
  decorator and signature stay on the clean method; only the
  body changes to a thin delegate. The two private helpers
  (_fan_out_clean_to_connected_peers, _active_build_for) are
  removed from the class entirely. Ruff auto-removed the
  now-unused imports (JobBuildSource, JobSource,
  PeerStatus, plus the moved _BUILD_PRODUCING_JOB_TYPES
  block).

Combined arc (#733 + #735 + #739 + #740 + #745 + #746 + this):
controller.py 2083 -> 877 lines (-58%). One module away from the
800-line soft cap.

All 3198 non-e2e tests pass; ruff + mypy clean. No
production-side behaviour change.
